Move consent state methods into mod
